+++ RxCocoa/iOS/UILabel+Rx.swift
@@ -28,6 +28,22 @@ extension Reactive where Base: UILabel {
             label.attributedText = text
         }
     }
+
+    public var rx_attributedText: AnyObserver<NSAttributedString> {
+        return AnyObserver { [weak self] event in
+            MainScheduler.ensureExecutingOnScheduler()
+
+            switch event {
+            case .Next(let value):
+                self?.attributedText = value
+            case .Error(let error):
+                bindingErrorToInterface(error)
+                break
+            case .Completed:
+                break
+            }
+        }
+    }
     
 }