+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+* Adds `rx_attributedText` to `UILabel`.
+
 ---
 
 ## Master